[S08P12A804-157] BE/Fix : Change Id type

+++ backend/src/main/java/com/ssafy/kkini/dto/MemoryUpdateFormDto.java
@@ -16,11 +16,11 @@ import java.util.List;
 @AllArgsConstructor
 public class MemoryUpdateFormDto {
 
-    private Long memoryId;
+    private int memoryId;
     @NotNull(message = "제목은 필수 입력 값입니다.")
     private String memoryTitle;
     private String memoryContent;
-    private Long userId;
+    private int userId;
     private List<MultipartFile> memoryImgFiles;
 
     public Memory toEntity(){
